Solve QT deprecation warning on float
Avoids two messages of the kind: DeprecationWarning: an integer is
required (got type float).  Implicit conversion to integers using
__int__ is deprecated, and may be removed in a future version of Python.

+++ ert_gui/tools/plot/color_chooser.py
@@ -22,7 +22,7 @@ class ColorBox(QFrame):
         painter = QPainter(self)
         rect = self.contentsRect()
         tile_count = 3
-        tile_size = rect.width() / tile_count
+        tile_size = int(rect.width() / tile_count)
         painter.save()
         painter.translate(rect.x(), rect.y())
